Will commence with the architecture described at https://github.com/serbanghita/Mobile-Detect/wiki/3.0-Architecture
