+++ lib/Data/browser.safari.version.json
@@ -1,0 +1,442 @@
+{
+  "0.8": {
+    "version": "48",
+    "codename": ""
+  },
+  "73": {
+    "version": "0.9",
+    "codename": ""
+  },
+  "85": {
+    "version": "1.0",
+    "codename": ""
+  },
+  "85.8.5": {
+    "version": "1.0.3",
+    "codename": ""
+  },
+  "100": {
+    "version": "1.1",
+    "codename": ""
+  },
+  "125": {
+    "version": "1.2",
+    "codename": ""
+  },
+  "312": {
+    "version": "1.3",
+    "codename": ""
+  },
+  "312.3": {
+    "version": "1.3.1",
+    "codename": ""
+  },
+  "312.5": {
+    "version": "1.3.2",
+    "codename": ""
+  },
+  "312.6": {
+    "version": "1.3.2",
+    "codename": ""
+  },
+  "412": {
+    "version": "2.0",
+    "codename": ""
+  },
+  "416.11": {
+    "version": "2.0.2",
+    "codename": ""
+  },
+  "419.3": {
+    "version": "2.0.4",
+    "codename": ""
+  },
+  "522.11": {
+    "version": "3.0",
+    "codename": ""
+  },
+  "522.12": {
+    "version": "3.0.2",
+    "codename": ""
+  },
+  "522.12.1": {
+    "version": "3.0.3",
+    "codename": ""
+  },
+  "523.10": {
+    "version": "3.0.4",
+    "codename": ""
+  },
+  "525.13": {
+    "version": "3.1",
+    "codename": ""
+  },
+  "525.17": {
+    "version": "3.1.1",
+    "codename": ""
+  },
+  "525.20": {
+    "version": "525.20",
+    "codename": "Safari Mobile"
+  },
+  "525.21": {
+    "version": "3.1.2",
+    "codename": ""
+  },
+  "525.26": {
+    "version": "3.2",
+    "codename": ""
+  },
+  "525.27": {
+    "version": "3.2.1",
+    "codename": ""
+  },
+  "525.28": {
+    "version": "3.2.3",
+    "codename": ""
+  },
+  "526.11.2": {
+    "version": "4.0 Beta",
+    "codename": ""
+  },
+  "528.16": {
+    "version": "528.16",
+    "codename": "Safari Mobile"
+  },
+  "528.17": {
+    "version": [
+      "4.0",
+      "4.0 Beta"
+    ],
+    "codename": ""
+  },
+  "530.17": {
+    "version": [
+      "4.0",
+      "4.0.1"
+    ],
+    "codename": ""
+  },
+  "530.18": {
+    "version": "4.0.1",
+    "codename": ""
+  },
+  "530.19": {
+    "version": "4.0.2",
+    "codename": ""
+  },
+  "531.9": {
+    "version": "4.0.3",
+    "codename": ""
+  },
+  "531.21.10": {
+    "version": "531.21.10",
+    "codename": "Safari Mobile for iOS 3.2"
+  },
+  "531.22.7": {
+    "version": "4.0.5",
+    "codename": ""
+  },
+  "533.16": {
+    "version": [
+      "4.1",
+      "5.0"
+    ],
+    "codename": ""
+  },
+  "533.17.8": {
+    "version": [
+      "4.1.1",
+      "5.0.1"
+    ],
+    "codename": ""
+  },
+  "533.17.9": {
+    "version": "5.0.2",
+    "codename": ""
+  },
+  "533.18.5": {
+    "version": [
+      "4.1.2",
+      "5.0.2"
+    ],
+    "codename": ""
+  },
+  "533.19.4": {
+    "version": [
+      "4.1.3",
+      "5.0.3"
+    ],
+    "codename": ""
+  },
+  "533.20.27": {
+    "version": "5.0.4",
+    "codename": ""
+  },
+  "533.21.1": {
+    "version": "5.0.5",
+    "codename": ""
+  },
+  "533.22.3": {
+    "version": "5.0.6",
+    "codename": ""
+  },
+  "534.6+": {
+    "version": "5.0.1",
+    "codename": ""
+  },
+  "534.8": {
+    "version": "5.0.2",
+    "codename": ""
+  },
+  "534.8+": {
+    "version": "5.0.2",
+    "codename": ""
+  },
+  "534.11": {
+    "version": "5.0.2",
+    "codename": ""
+  },
+  "534.11+": {
+    "version": "5.0.2",
+    "codename": ""
+  },
+  "534.48.3": {
+    "version": "5.1",
+    "codename": ""
+  },
+  "534.51.22": {
+    "version": "5.1.1",
+    "codename": ""
+  },
+  "534.52.7": {
+    "version": "5.1.2",
+    "codename": ""
+  },
+  "534.53.10": {
+    "version": "5.1.3",
+    "codename": ""
+  },
+  "534.54.16": {
+    "version": "5.1.4",
+    "codename": ""
+  },
+  "534.55.3": {
+    "version": "5.1.5",
+    "codename": ""
+  },
+  "534.56.5": {
+    "version": "5.1.6",
+    "codename": ""
+  },
+  "534.57.2": {
+    "version": "5.1.7",
+    "codename": ""
+  },
+  "534.58.2": {
+    "version": "5.1.8",
+    "codename": ""
+  },
+  "534.59.8": {
+    "version": "5.1.9",
+    "codename": ""
+  },
+  "534.59.10": {
+    "version": "5.1.10",
+    "codename": ""
+  },
+  "536.2": {
+    "version": "5.1.2",
+    "codename": ""
+  },
+  "536.2+": {
+    "version": "5.1.2",
+    "codename": ""
+  },
+  "536.25": {
+    "version": "6.0",
+    "codename": ""
+  },
+  "536.26": {
+    "version": "6.0.1",
+    "codename": ""
+  },
+  "536.26.17": {
+    "version": "6.0.2",
+    "codename": ""
+  },
+  "536.28.10": {
+    "version": "6.0.3",
+    "codename": ""
+  },
+  "536.29.13": {
+    "version": "6.0.4",
+    "codename": ""
+  },
+  "536.30.1": {
+    "version": "6.0.5",
+    "codename": ""
+  },
+  "537.1+": {
+    "version": "5.1.5",
+    "codename": ""
+  },
+  "537.10": {
+    "version": "5.1",
+    "codename": ""
+  },
+  "537.10+": {
+    "version": "5.1",
+    "codename": ""
+  },
+  "537.43.58": {
+    "version": "6.1",
+    "codename": ""
+  },
+  "537.51.1": {
+    "version": "",
+    "codename": ""
+  },
+  "537.73.11": {
+    "version": [
+      "6.1.1",
+      "7.0.1"
+    ],
+    "codename": ""
+  },
+  "537.85.17": {
+    "version": [
+      "6.2.8",
+      "7.1.8"
+    ],
+    "codename": ""
+  },
+  "537.71": {
+    "version": "7.0",
+    "codename": ""
+  },
+  "537.75.14": {
+    "version": "7.0.3",
+    "codename": ""
+  },
+  "537.76.4": {
+    "version": "7.0.4",
+    "codename": ""
+  },
+  "537.77.4": {
+    "version": "7.0.5",
+    "codename": ""
+  },
+  "537.78.2": {
+    "version": "7.0.6",
+    "codename": ""
+  },
+  "538.35.8": {
+    "version": "8.0",
+    "codename": ""
+  },
+  "600.6.3": {
+    "version": "8.0.6",
+    "codename": ""
+  },
+  "600.7.12": {
+    "version": "8.0.7",
+    "codename": ""
+  },
+  "522.11.3": {
+    "version": "3.0",
+    "codename": ""
+  },
+  "522.12.2": {
+    "version": "3.0.1",
+    "codename": ""
+  },
+  "522.13.1": {
+    "version": "3.0.2",
+    "codename": ""
+  },
+  "522.15.5": {
+    "version": "3.0.3",
+    "codename": ""
+  },
+  "523.12.9": {
+    "version": "3.0.4",
+    "codename": ""
+  },
+  "523.13": {
+    "version": "3.0.4",
+    "codename": ""
+  },
+  "523.15": {
+    "version": "3.0.4",
+    "codename": ""
+  },
+  "525.26.13": {
+    "version": "3.2",
+    "codename": ""
+  },
+  "525.27.1": {
+    "version": "3.2.1",
+    "codename": ""
+  },
+  "525.28.1": {
+    "version": "3.2.2",
+    "codename": ""
+  },
+  "525.29.1": {
+    "version": "3.2.3",
+    "codename": ""
+  },
+  "526.12.2": {
+    "version": "4.0",
+    "codename": ""
+  },
+  "528.1.1": {
+    "version": "4.0",
+    "codename": ""
+  },
+  "530.19.1": {
+    "version": "4.0.2",
+    "codename": ""
+  },
+  "531.9.1": {
+    "version": "4.0.3",
+    "codename": ""
+  },
+  "534.30": {
+    "version": "5.1",
+    "codename": ""
+  },
+  "534.50": {
+    "version": "5.1",
+    "codename": ""
+  },
+  "600.1.4": {
+    "version": "600.1.4",
+    "codename": "Safari Mobile for iOS 8.x"
+  },
+  "9537.53": {
+    "version": "9537.53",
+    "codename": "Safari Mobile for iOS 7.x"
+  },
+  "8536.25": {
+    "version": "8536.25",
+    "codename": "Safari Mobile for iOS 6.x"
+  },
+  "7534.48.3": {
+    "version": "7534.48.3",
+    "codename": "Safari Mobile for iOS 5.x"
+  },
+  "6533.18.5": {
+    "version": "6533.18.5",
+    "codename": "Safari Mobile 5.0.2"
+  },
+  "6531.22.7": {
+    "version": "6531.22.7",
+    "codename": "Safari Mobile"
+  },
+  "525.18.1": {
+    "version": "525.18.1",
+    "codename": "Safari Mobile"
+  }
+}
