+++ lib/Data/operatingsystem.iossdk.version.json
@@ -1,0 +1,149 @@
+{
+  "5A147p": "iPhone OS 1.2 Beta 1",
+  "5A225c": "iPhone OS 2.0 Beta 2",
+  "5A240d": "iPhone OS 2.0 Beta 3",
+  "5A258f": "iPhone OS 2.0 Beta 4",
+  "5A274d": "iPhone OS 2.0 Beta 5",
+  "5A292g": "iPhone OS 2.0 Beta 6",
+  "5A308": "iPhone OS 2.0 Beta 6",
+  "5A347": "iPhone OS 2.0 Final",
+  "5F90": "iPhone OS 2.1 Beta",
+  "5F97": "iPhone OS 2.1 Beta 2",
+  "5F108": "iPhone OS 2.1 Beta 3",
+  "5F116": "iPhone OS 2.1 Beta 4",
+  "5F136": "iPhone OS 2.1 Final",
+  "5G29": "iPhone OS 2.2 Beta",
+  "5G53": "iPhone OS 2.2 Beta 2",
+  "5G77": "iPhone OS 2.2 Final",
+  "7A238j": "iPhone OS 3.0 Beta 1",
+  "7A259g": "iPhone OS 3.0 Beta 2",
+  "7A280f": "iPhone OS 3.0 Beta 3",
+  "7A300g": "iPhone OS 3.0 Beta 4",
+  "7A312g": "iPhone OS 3.0 Beta 5",
+  "5A331": "iPhone OS 3.0 Golden Master",
+  "5A345": "iPhone OS 3.0 Final",
+  "7C97d": "iPhone OS 3.1 Beta 1",
+  "7C106c": "iPhone OS 3.1 Beta 2",
+  "7C116a": "iPhone OS 3.1 Beta 3",
+  "7C144": "iPhone OS 3.1 Final",
+  "7B314": "iOS 3.2 Beta 1",
+  "7B320c": "iOS 3.2 Beta 2",
+  "7B334b": "iOS 3.2 Beta 3",
+  "7B348b": "iOS 3.2 Beta 4",
+  "7B360": "iOS 3.2 Beta 5",
+  "7B367": "iOS 3.2 Final",
+  "8A230m": "iOS 4.0 Beta 1",
+  "8A248c": "iOS 4.0 Beta 2",
+  "8A260b": "iOS 4.0 Beta 3",
+  "8A293": "iOS 4.0 Golden Master",
+  "8A331": "iOS 4.0 Final",
+  "8B5080c": "iOS 4.1 Beta 1",
+  "8B5091b": "iOS 4.1 Beta 2",
+  "8B5097d": "iOS 4.1 Beta 3",
+  "8B117": "iOS 4.1 Final",
+  "8C5091e": "iOS 4.2 Beta 1",
+  "8C5101c": "iOS 4.2 Beta 2",
+  "8C5115c": "iOS 4.2 Beta 3",
+  "8C134": "iOS 4.2 Golden Master",
+  "8C148": "iOS 4.2 Golden Master 2",
+  "8F5148b": "iOS 4.3 Beta 1",
+  "8F5153d": "iOS 4.3 Beta 2",
+  "8F5166b": "iOS 4.3 Beta 3",
+  "8F190": "iOS 4.3 Final",
+  "9a5220p": "iOS 5.0 beta 1",
+  "9A524Bd": "iOS 5.0 beta 2",
+  "9A5259f": "iOS 5.0 beta 3",
+  "9A5274d": "iOS 5.0 beta 4",
+  "9A5288d": "iOS 5.0 beta 5",
+  "9A5302b": "iOS 5.0 beta 6",
+  "9A5313e": "iOS 5.0 beta 7",
+  "9A334": "iOS 5.0 Final",
+  "9A402": "iOS 5.0.1 beta 1",
+  "9A405": "iOS 5.0.1 Final",
+  "9B5117b": "iOS 5.1 beta 1",
+  "9B5127c": "iOS 5.1 beta 2",
+  "9B5141a": "iOS 5.1 beta 3",
+  "9B176": "iOS 5.1 Final",
+  "9B179": "iOS 5.1 Final",
+  "10A5316k": "iOS 6.0 beta 1",
+  "10A5338d": "iOS 6.0 beta 2",
+  "10A5355d": "iOS 6.0 beta 3",
+  "10A5376e": "iOS 6.0 beta 4",
+  "10A403": "iOS 6.0 Golden Master",
+  "10A405": "iOS 6.0 Final",
+  "10A406": "iOS 6.0 Final",
+  "10B5095f": "iOS 6.1 beta 1",
+  "10B5105c": "iOS 6.1 beta 2",
+  "10B5117b": "iOS 6.1 beta 3",
+  "10B5126b": "iOS 6.1 beta 4",
+  "10B141": "iOS 6.1 Final",
+  "10B142": "iOS 6.1 Final",
+  "10B143": "iOS 6.1 Final",
+  "10B144": "iOS 6.1 Final",
+  "11A4372q": "iOS 7.0 beta 1",
+  "11A4400f": "iOS 7.0 beta 2",
+  "11A4414e": "iOS 7.0 beta 3",
+  "11A4435d": "iOS 7.0 beta 4",
+  "11A4449a": "iOS 7.0 beta 5",
+  "11A4449d": "iOS 7.0 beta 6",
+  "11A465": "iOS 7.0 Final",
+  "11D5099e": "iOS 7.1 beta 1",
+  "11D5115d": "iOS 7.1 beta 2",
+  "11D5127c": "iOS 7.1 beta 3",
+  "11D5134c": "iOS 7.1 beta 4",
+  "11D5145e": "iOS 7.1 beta 5",
+  "11D167": "iOS 7.1 Final",
+  "12A4265u": "iOS 8.0 beta 1",
+  "12A4297e": "iOS 8.0 beta 2",
+  "12A4318c": "iOS 8.0 beta 3",
+  "12A4331d": "iOS 8.0 beta 4",
+  "12A4345d": "iOS 8.0 beta 5",
+  "12A365": "iOS 8.0 Final",
+  "12B401": "iOS 8.1 beta 1",
+  "12B407": "iOS 8.1 beta 2",
+  "12B410": "iOS 8.1 Final",
+  "12B411": "iOS 8.1 Final",
+  "12B432": "iOS 8.1.1 beta 1",
+  "12B435": "iOS 8.1.1 Final",
+  "12B436": "iOS 8.2 beta 1",
+  "12D445d": "iOS 8.2 beta 2",
+  "12D5452a": "iOS 8.2 beta 3",
+  "12D5461b": "iOS 8.2 beta 4",
+  "12D5480a": "iOS 8.2 beta 5",
+  "12D508": "iOS 8.2 Final",
+  "12F5027d": "iOS 8.3 beta 1",
+  "12F5037c": "iOS 8.3 beta 2",
+  "12F5047f": "iOS 8.3 beta 3",
+  "12F61": "iOS 8.3 beta 4",
+  "12F69": "iOS 8.3 Final",
+  "12F70": "iOS 8.3 Final",
+  "12H4074d": "iOS 8.4 beta 1",
+  "12H4086d": "iOS 8.4 beta 2",
+  "12H4098c": "iOS 8.4 beta 3",
+  "12H4125a": "iOS 8.4 beta 4",
+  "12H143": "iOS 8.4 Final",
+  "12H304": "iOS 8.4.1 beta 1",
+  "12H318": "iOS 8.4.1 beta 2",
+  "12H321": "iOS 8.4.1 Final",
+  "13A4254v": "iOS 9.0 beta 1",
+  "13A4280e": "iOS 9.0 beta 2",
+  "13A4293g": "iOS 9.0 beta 3",
+  "13A4305g": "iOS 9.0 beta 4",
+  "13A4325c": "iOS 9.0 beta 5",
+  "13A340": "iOS 9.0 Final",
+  "13A404": "iOS 9.0.1",
+  "13A452": "iOS 9.0.2",
+  "13B5110e": "iOS 9.1 beta 1",
+  "13B5119e": "iOS 9.1 beta 2",
+  "13B5130b": "iOS 9.1 beta 3",
+  "13B136": "iOS 9.1 beta 4",
+  "13B137": "iOS 9.1 beta 5",
+  "13B139": "iOS 9.1 beta 5",
+  "13B143": "iOS 9.1 Final",
+  "13C5055d": "iOS 9.2 beta 1",
+  "13C5060d": "iOS 9.2 beta 2",
+  "13C71": "iOS 9.2 beta 3",
+  "13C5075": "iOS 9.2 beta 4",
+  "13C75": "iOS 9.2 Final",
+  "13D11": "iOS 9.2.1 beta 1"
+}
