--- lib/Repository/Data/test.php
@@ -1,151 +1,0 @@
-<?php
-$data = [
-    'iPhone' => [
-        'vendor' => 'Apple',
-
-        'matchIdentity' => [
-            [
-                'context' => null,
-                'match' => ['\b(iPhone|iPod)\b'],
-                'matchType' => 'regex'
-            ]
-        ],
-        'matchModelAndVersion' => [
-            [
-                'context' => null,
-                'match' => [
-                    '(?<model>iP[a-z]+).*CPU[a-z ]+(?<version>[0-9\._-]+)',
-                    '(?<model>iP[a-z]+).*CPU[a-z ]+(?<version>[0-9\._-]+)',
-                    '(?<model>iP[a-z]+); U; (.* OS (?<version>[0-9\._-]+))?',
-                    '(?<model>iP[a-z]+); iPhone OS/(?<version>[0-9\._-]+)',
-                    '(?<model>iP[a-z]+); iOS (?<version>[0-9\._-]+);',
-                ],
-                'matchType' => 'regex'
-            ]
-        ],
-    ],
-
-    'Nexus' => [
-        'vendor' => 'Google',
-
-        'matchIdentity' => [
-            [
-                'context' => null,
-                'match' => ['Nexus One|Nexus S|Galaxy.*Nexus|Android.*Nexus.*Mobile|Nexus 4|Nexus 5|Nexus 6'],
-                'matchType' => 'regex'
-            ]
-        ],
-        'matchModelAndVersion' => [
-            [
-                'context' => null,
-                'match' => [
-                    'Google (?<model>[a-zA-Z]+ (?<version>[a-zA-Z0-9]+))',
-                    '(?<model>Nexus (?<version>[a-zA-Z0-9]+)) Build',
-                    '(?<model>[a-zA-Z0-9]+ Nexus)'
-                ],
-                'matchType' => 'regex'
-            ]
-        ],
-    ],
-
-    'Samsung' => [
-        'vendor' => 'Samsung',
-
-        'matchIdentity' => [
-            [
-                'context' => null,
-                'match' => [
-                    'Samsung|\bGalaxy\b|SGH-I337|BGT-S5230|GT-B2100|GT-B2700|GT-B2710|GT-B3210|GT-B3310|GT-B3410|GT-B3730|GT-B3740|GT-B5510|GT-B5512|GT-B5722|GT-B6520|GT-B7300|GT-B7320',
-                    '|GT-B7330|GT-B7350|GT-B7510|GT-B7722|GT-B7800|GT-C3010|GT-C3011|GT-C3060|GT-C3200|GT-C3212|GT-C3212I|GT-C3262|GT-C3222|GT-C3300|GT-C3300K|GT-C3303|GT-C3303K',
-                    '|GT-C3310|GT-C3322|GT-C3330|GT-C3350|GT-C3500|GT-C3510|GT-C3530|GT-C3630|GT-C3780|GT-C5010|GT-C5212|GT-C6620|GT-C6625|GT-C6712|GT-E1050|GT-E1070|GT-E1075',
-                    '|GT-E1080|GT-E1081|GT-E1085|GT-E1087|GT-E1100|GT-E1107|GT-E1110|GT-E1120|GT-E1125|GT-E1130|GT-E1160|GT-E1170|GT-E1175|GT-E1180|GT-E1182|GT-E1200|GT-E1210',
-                    '|GT-E1225|GT-E1230|GT-E1390|GT-E2100|GT-E2120|GT-E2121|GT-E2152|GT-E2220|GT-E2222|GT-E2230|GT-E2232|GT-E2250|GT-E2370|GT-E2550|GT-E2652|GT-E3210|GT-E3213',
-                    '|GT-I5500|GT-I5503|GT-I5700|GT-I5800|GT-I5801|GT-I6410|GT-I6420|GT-I7110|GT-I7410|GT-I7500|GT-I8000|GT-I8150|GT-I8160|GT-I8190|GT-I8320|GT-I8330|GT-I8350',
-                    '|GT-I8530|GT-I8700|GT-I8703|GT-I8910|GT-I9000|GT-I9001|GT-I9003|GT-I9010|GT-I9020|GT-I9023|GT-I9070|GT-I9082|GT-I9100|GT-I9103|GT-I9220|GT-I9250|GT-I9300',
-                    '|GT-I9305|GT-I9500|GT-I9505|GT-M3510|GT-M5650|GT-M7500|GT-M7600|GT-M7603|GT-M8800|GT-M8910|GT-N7000|GT-S3110|GT-S3310|GT-S3350|GT-S3353|GT-S3370|GT-S3650',
-                    '|GT-S3653|GT-S3770|GT-S3850|GT-S5210|GT-S5220|GT-S5229|GT-S5230|GT-S5233|GT-S5250|GT-S5253|GT-S5260|GT-S5263|GT-S5270|GT-S5300|GT-S5330|GT-S5350|GT-S5360',
-                    '|GT-S5363|GT-S5369|GT-S5380|GT-S5380D|GT-S5560|GT-S5570|GT-S5600|GT-S5603|GT-S5610|GT-S5620|GT-S5660|GT-S5670|GT-S5690|GT-S5750|GT-S5780|GT-S5830|GT-S5839',
-                    '|GT-S6102|GT-S6500|GT-S7070|GT-S7200|GT-S7220|GT-S7230|GT-S7233|GT-S7250|GT-S7500|GT-S7530|GT-S7550|GT-S7562|GT-S7710|GT-S8000|GT-S8003|GT-S8500|GT-S8530',
-                    '|GT-S8600|SCH-A310|SCH-A530|SCH-A570|SCH-A610|SCH-A630|SCH-A650|SCH-A790|SCH-A795|SCH-A850|SCH-A870|SCH-A890|SCH-A930|SCH-A950|SCH-A970|SCH-A990|SCH-I100',
-                    '|SCH-I110|SCH-I400|SCH-I405|SCH-I500|SCH-I510|SCH-I515|SCH-I600|SCH-I730|SCH-I760|SCH-I770|SCH-I830|SCH-I910|SCH-I920|SCH-I959|SCH-LC11|SCH-N150|SCH-N300',
-                    '|SCH-R100|SCH-R300|SCH-R351|SCH-R400|SCH-R410|SCH-T300|SCH-U310|SCH-U320|SCH-U350|SCH-U360|SCH-U365|SCH-U370|SCH-U380|SCH-U410|SCH-U430|SCH-U450|SCH-U460',
-                    '|SCH-U470|SCH-U490|SCH-U540|SCH-U550|SCH-U620|SCH-U640|SCH-U650|SCH-U660|SCH-U700|SCH-U740|SCH-U750|SCH-U810|SCH-U820|SCH-U900|SCH-U940|SCH-U960|SCS-26UC',
-                    '|SGH-A107|SGH-A117|SGH-A127|SGH-A137|SGH-A157|SGH-A167|SGH-A177|SGH-A187|SGH-A197|SGH-A227|SGH-A237|SGH-A257|SGH-A437|SGH-A517|SGH-A597|SGH-A637|SGH-A657',
-                    '|SGH-A667|SGH-A687|SGH-A697|SGH-A707|SGH-A717|SGH-A727|SGH-A737|SGH-A747|SGH-A767|SGH-A777|SGH-A797|SGH-A817|SGH-A827|SGH-A837|SGH-A847|SGH-A867|SGH-A877',
-                    '|SGH-A887|SGH-A897|SGH-A927|SGH-B100|SGH-B130|SGH-B200|SGH-B220|SGH-C100|SGH-C110|SGH-C120|SGH-C130|SGH-C140|SGH-C160|SGH-C170|SGH-C180|SGH-C200|SGH-C207',
-                    '|SGH-C210|SGH-C225|SGH-C230|SGH-C417|SGH-C450|SGH-D307|SGH-D347|SGH-D357|SGH-D407|SGH-D415|SGH-D780|SGH-D807|SGH-D980|SGH-E105|SGH-E200|SGH-E315|SGH-E316',
-                    '|SGH-E317|SGH-E335|SGH-E590|SGH-E635|SGH-E715|SGH-E890|SGH-F300|SGH-F480|SGH-I200|SGH-I300|SGH-I320|SGH-I550|SGH-I577|SGH-I600|SGH-I607|SGH-I617|SGH-I627',
-                    '|SGH-I637|SGH-I677|SGH-I700|SGH-I717|SGH-I727|SGH-I747|SGH-i747M|SGH-I777|SGH-I780|SGH-I827|SGH-I847|SGH-I857|SGH-I896|SGH-I897|SGH-I900|SGH-I907|SGH-I917|SGH-I927',
-                    '|SGH-I937|SGH-I997|SGH-J150|SGH-J200|SGH-L170|SGH-L700|SGH-M110|SGH-M150|SGH-M200|SGH-N105|SGH-N500|SGH-N600|SGH-N620|SGH-N625|SGH-N700|SGH-N710|SGH-P107',
-                    '|SGH-P207|SGH-P300|SGH-P310|SGH-P520|SGH-P735|SGH-P777|SGH-Q105|SGH-R210|SGH-R220|SGH-R225|SGH-S105|SGH-S307|SGH-T109|SGH-T119|SGH-T139|SGH-T209|SGH-T219',
-                    '|SGH-T229|SGH-T239|SGH-T249|SGH-T259|SGH-T309|SGH-T319|SGH-T329|SGH-T339|SGH-T349|SGH-T359|SGH-T369|SGH-T379|SGH-T409|SGH-T429|SGH-T439|SGH-T459|SGH-T469',
-                    '|SGH-T479|SGH-T499|SGH-T509|SGH-T519|SGH-T539|SGH-T559|SGH-T589|SGH-T609|SGH-T619|SGH-T629|SGH-T639|SGH-T659|SGH-T669|SGH-T679|SGH-T709|SGH-T719|SGH-T729',
-                    '|SGH-T739|SGH-T746|SGH-T749|SGH-T759|SGH-T769|SGH-T809|SGH-T819|SGH-T839|SGH-T919|SGH-T929|SGH-T939|SGH-T959|SGH-T989|SGH-U100|SGH-U200|SGH-U800|SGH-V205',
-                    '|SGH-V206|SGH-X100|SGH-X105|SGH-X120|SGH-X140|SGH-X426|SGH-X427|SGH-X475|SGH-X495|SGH-X497|SGH-X507|SGH-X600|SGH-X610|SGH-X620|SGH-X630|SGH-X700|SGH-X820',
-                    '|SGH-X890|SGH-Z130|SGH-Z150|SGH-Z170|SGH-ZX10|SGH-ZX20|SHW-M110|SPH-A120|SPH-A400|SPH-A420|SPH-A460|SPH-A500|SPH-A560|SPH-A600|SPH-A620|SPH-A660|SPH-A700',
-                    '|SPH-A740|SPH-A760|SPH-A790|SPH-A800|SPH-A820|SPH-A840|SPH-A880|SPH-A900|SPH-A940|SPH-A960|SPH-D600|SPH-D700|SPH-D710|SPH-D720|SPH-I300|SPH-I325|SPH-I330',
-                    '|SPH-I350|SPH-I500|SPH-I600|SPH-I700|SPH-L700|SPH-M100|SPH-M220|SPH-M240|SPH-M300|SPH-M305|SPH-M320|SPH-M330|SPH-M350|SPH-M360|SPH-M370|SPH-M380|SPH-M510',
-                    '|SPH-M540|SPH-M550|SPH-M560|SPH-M570|SPH-M580|SPH-M610|SPH-M620|SPH-M630|SPH-M800|SPH-M810|SPH-M850|SPH-M900|SPH-M910|SPH-M920|SPH-M930|SPH-N100|SPH-N200',
-                    '|SPH-N240|SPH-N300|SPH-N400|SPH-Z400|SWC-E100|SCH-i909|GT-N7100|GT-N7105|SCH-I535|SM-N900A|SGH-I317|SGH-T999L|GT-S5360B|GT-I8262|GT-S6802|GT-S6312|GT-S6310',
-                    '|GT-S5312|GT-S5310|GT-I9105|GT-I8510|GT-S6790N|SM-G7105|SM-N9005|GT-S5301|GT-I9295|GT-I9195|SM-C101|GT-S7392|GT-S7560|GT-B7610|GT-I5510|GT-S7582|GT-S7530E',
-                    '|GT-I8750|SM-G9006V|SM-G9008V|SM-G9009D|SM-G900A|SM-G900D|SM-G900F|SM-G900H|SM-G900I|SM-G900J|SM-G900K|SM-G900L|SM-G900M|SM-G900P|SM-G900R4|SM-G900S|SM-G900T',
-                    '|SM-G900V|SM-G900W8|SHV-E160K|GT-S5367|SM-G9250'
-                ]
-            ]
-        ],
-        'matchModelAndVersion' => [
-            [
-                'context' => null,
-                'match' => [
-                    'SAMSUNG (?<model>[a-zA-Z0-9-]+)',
-                    '; SAMSUNG; (?<model>[a-zA-Z0-9- ]+)',
-                    'SAMSUNG-(?<model>[a-zA-Z0-9-]+)',
-                    '; (?<model>[a-zA-Z0-9-]+) Build',
-                    '; (?<model>[a-zA-Z0-9-]+)\) like Gecko',
-                    '; (?<model>[a-zA-Z0-9-]+)/[a-zA-Z0-9-.]+ Build',
-                    //'[a-z]{2}-[a-z]{2}; (?<model>[a-zA-Z0-9-]+) Build',
-                    '[a-z]{2}-[a-z]{2}; (?<model>[a-z]{2}-[a-z0-9]+);'
-                ],
-                'matchType' => 'regex'
-            ]
-        ]
-    ],
-
-    'BlackBerry' => [
-        'vendor' => 'BlackBerry',
-
-        'matchIdentity' => [
-            [
-                'context' => null,
-                'match' => ['BlackBerry|\bBB10\b|rim[0-9]+']
-            ]
-        ],
-        'matchModelAndVersion' => [
-            [
-                'context' => null,
-                'match' => [
-                    'BlackBerry (?<model>[a-zA-Z0-9-]+);',
-                    'BlackBerry(?<model>[a-zA-Z0-9-]+)'
-                ]
-            ]
-        ]
-    ],
-
-    'Acer' => [
-        'vendor' => 'Acer',
-
-        'matchIdentity' => false,
-
-        'matchModelAndVersion' => [
-            [
-                'context' => null,
-                'match' => [
-                    '; (?<model>[a-zA-Z0-9-]+) Build',
-                    '; Acer; (?<model>[a-zA-Z0-9-]+)'
-                ]
-            ]
-        ]
-    ]
-];
-
-echo json_encode($data, JSON_PRETTY_PRINT);
