+++ lib/Repository/OperatingSystem/OperatingSystemInterface.php
@@ -1,0 +1,9 @@
+<?php
+namespace MobileDetect\Repository\OperatingSystem;
+
+use MobileDetect\Repository\Browser\BrowserInterface;
+
+interface OperatingSystemInterface extends BrowserInterface
+{
+    
+}
